# openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateMatchEndorsementDto": {
        "properties": {
          "strengths": {
            "example": [
              "SMASH",
              "DEFENSA"
            ],
            "items": {
              "enum": [
                "SMASH",
                "BANDEJA",
                "VIBORA",
                "VOLEA",
                "GLOBO",
                "DEFENSA",
                "RESILIENCIA",
                "TACTICA",
                "COMUNICACION",
                "VELOCIDAD",
                "PRECISION"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "type": "array"
          },
          "toUserId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "toUserId",
          "strengths"
        ],
        "type": "object"
      },
      "CreateMatchEndorsementResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "fromUserId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "matchId": {
            "format": "uuid",
            "type": "string"
          },
          "strengths": {
            "items": {
              "enum": [
                "SMASH",
                "BANDEJA",
                "VIBORA",
                "VOLEA",
                "GLOBO",
                "DEFENSA",
                "RESILIENCIA",
                "TACTICA",
                "COMUNICACION",
                "VELOCIDAD",
                "PRECISION"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "type": "array"
          },
          "toUserId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "matchId",
          "fromUserId",
          "toUserId",
          "strengths",
          "createdAt"
        ],
        "type": "object"
      },
...
      "DisputeMatchDto": {
        "properties": {},
        "type": "object"
      },
...
      "MatchRankingImpactResponseDto": {
        "properties": {
          "categoryAfter": {
            "example": 6,
            "nullable": true,
            "type": "number"
          },
          "categoryBefore": {
            "example": 6,
            "nullable": true,
            "type": "number"
          },
          "eloAfter": {
            "example": 1470,
            "nullable": true,
            "type": "number"
          },
          "eloBefore": {
            "example": 1450,
            "nullable": true,
            "type": "number"
          },
          "eloDelta": {
            "example": 20,
            "type": "number"
          },
          "impactRanking": {
            "example": true,
            "type": "boolean"
          },
          "matchId": {
            "format": "uuid",
            "type": "string"
          },
          "positionAfter": {
            "example": 14,
            "nullable": true,
            "type": "number"
          },
          "positionBefore": {
            "example": 16,
            "nullable": true,
            "type": "number"
          },
          "positionDelta": {
            "example": 2,
            "type": "number"
          },
          "result": {
            "enum": [
              "WIN",
              "LOSS",
              "DRAW"
            ],
            "example": "WIN",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/MatchRankingImpactSummaryDto"
          },
          "viewerUserId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "matchId",
          "viewerUserId",
          "result",
          "eloDelta",
          "positionDelta",
          "impactRanking",
          "summary"
        ],
        "type": "object"
      },
      "MatchRankingImpactSummaryDto": {
        "properties": {
          "subtitle": {
            "example": "+20 ELO despues de este partido",
            "type": "string"
          },
          "title": {
            "example": "Ganaste y subiste 2 posiciones",
            "type": "string"
          }
        },
        "required": [
          "title",
          "subtitle"
        ],
        "type": "object"
      },
...
      "MyPendingConfirmationItemDto": {
        "properties": {
          "cta": {
            "$ref": "#/components/schemas/PendingConfirmationCtaDto"
          },
          "id": {
            "type": "string"
          },
          "leagueId": {
            "nullable": true,
            "type": "string"
          },
          "leagueName": {
            "nullable": true,
            "type": "string"
          },
          "matchId": {
            "type": "string"
          },
          "opponentAvatarUrl": {
            "nullable": true,
            "type": "string"
          },
          "opponentName": {
            "description": "Stable opponent display label. Never empty.",
            "type": "string"
          },
          "playedAt": {
            "description": "ISO 8601 timestamp when the match was played",
            "type": "string"
          },
          "score": {
            "description": "Compact score label, ex: 6-4 6-2",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING_CONFIRMATION"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "matchId",
          "status",
          "opponentName",
          "cta"
        ],
        "type": "object"
      },
      "MyPendingConfirmationsResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MyPendingConfirmationItemDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "Opaque cursor for the next page; null when no more items",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "PendingConfirmationCtaDto": {
        "properties": {
          "href": {
            "type": "string"
          },
          "primary": {
            "enum": [
              "Confirmar",
              "Ver"
            ],
            "type": "string"
          }
        },
        "required": [
          "primary"
        ],
        "type": "object"
      },
...
      "RejectMatchDto": {
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ReportMatchDto": {
        "properties": {
          "challengeId": {
            "type": "string"
          },
          "leagueId": {
            "format": "uuid",
            "type": "string"
          },
          "playedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/ReportSetDto"
            },
            "maxItems": 3,
            "minItems": 2,
            "type": "array"
          }
        },
        "required": [
          "challengeId",
          "sets"
        ],
        "type": "object"
      },
      "ReportSetDto": {
        "properties": {
          "a": {
            "example": 6,
            "maximum": 7,
            "minimum": 0,
            "type": "number"
          },
          "b": {
            "example": 4,
            "maximum": 7,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "a",
          "b"
        ],
        "type": "object"
      },
...
      "ResolveDisputeDto": {
        "properties": {},
        "type": "object"
      },
...
      "StrengthSummaryItemDto": {
        "properties": {
          "count": {
            "type": "number"
          },
          "percent": {
            "description": "Integer percentage 0..100",
            "type": "number"
          },
          "strength": {
            "allOf": [
              {
                "$ref": "#/components/schemas/player_strength_enum"
              }
            ]
          }
        },
        "required": [
          "strength",
          "count",
          "percent"
        ],
        "type": "object"
      },
      "StrengthSummaryResponseDto": {
        "properties": {
          "days": {
            "type": "number"
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/StrengthSummaryItemDto"
            },
            "type": "array"
          },
          "totalVotes": {
            "type": "number"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "days",
          "totalVotes",
          "strengths"
        ],
        "type": "object"
      },
...
      "player_strength_enum": {
        "enum": [
          "SMASH",
          "BANDEJA",
          "VIBORA",
          "VOLEA",
          "GLOBO",
          "DEFENSA",
          "RESILIENCIA",
          "TACTICA",
          "COMUNICACION",
          "VELOCIDAD",
          "PRECISION"
        ],
        "type": "string"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Backend API contract for PadelPoint",
    "title": "PadelPoint API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/matches": {
      "get": {
        "operationId": "MatchesController_getByChallenge",
        "parameters": [
          {
            "in": "query",
            "name": "challengeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      },
      "post": {
        "operationId": "MatchesController_report",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportMatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/me": {
      "get": {
        "operationId": "MatchesController_getMyMatches",
        "parameters": [
          {
            "description": "Compatibility flag. Use legacy=1 to return a plain array instead of the canonical wrapper.",
            "in": "query",
            "name": "legacy",
            "required": false,
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "properties": {
                        "items": {
                          "items": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "nextCursor": {
                          "nullable": true,
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    {
                      "items": {
                        "additionalProperties": true,
                        "type": "object"
                      },
                      "type": "array"
                    }
                  ]
                }
              }
            },
            "description": "Canonical wrapper by default. Legacy clients can request a plain array with legacy=1."
          }
        },
        "summary": "List current user matches. Canonical contract returns { items, nextCursor }.",
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/me/pending-confirmations": {
      "get": {
        "operationId": "MatchesController_getPendingConfirmations",
        "parameters": [
          {
            "description": "Opaque pagination cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyPendingConfirmationsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}": {
      "get": {
        "operationId": "MatchesController_getById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/admin-confirm": {
      "patch": {
        "operationId": "MatchesController_adminConfirm",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/confirm": {
      "patch": {
        "operationId": "MatchesController_confirm",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/dispute": {
      "post": {
        "operationId": "MatchesController_dispute",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/ranking-impact": {
      "get": {
        "description": "Returns the competitive impact of a confirmed match from the authenticated participant perspective. Position fields are only populated when exact snapshot context exists; otherwise they are null.",
        "operationId": "MatchesController_getRankingImpact",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRankingImpactResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get ranking impact for the authenticated player",
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/reject": {
      "patch": {
        "operationId": "MatchesController_reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/resolve": {
      "post": {
        "operationId": "MatchesController_resolve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{id}/resolve-confirm-as-is": {
      "post": {
        "operationId": "MatchesController_resolveConfirmAsIs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/matches/{matchId}/endorsements": {
      "post": {
        "description": "Lets an authenticated participant optionally endorse one rival from a confirmed match with up to 2 strengths. Each rival can only be endorsed once per match.",
        "operationId": "MatchEndorsementsController_create",
        "parameters": [
          {
            "in": "path",
            "name": "matchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchEndorsementDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMatchEndorsementResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create post-match endorsement for a rival",
        "tags": [
          "endorsements"
        ]
      }
    },
...
    "/players/{userId}/strengths": {
      "get": {
        "description": "Returns aggregated strengths inferred from post-match endorsements. This data feeds compact scouting and competitive profile surfaces.",
        "operationId": "PlayerStrengthsController_getStrengths",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Rolling window in days",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 90,
              "maximum": 3650,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrengthSummaryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get player strengths from endorsements",
        "tags": [
          "endorsements"
        ]
      }
    },
    "/players/{userId}/strengths/summary": {
      "get": {
        "description": "Backward-compatible alias for GET /players/:userId/strengths.",
        "operationId": "PlayerStrengthsController_getSummary",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Rolling window in days",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 90,
              "maximum": 3650,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrengthSummaryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get player strengths summary",
        "tags": [
          "endorsements"
        ]
      }
    },
...
  },
  "servers": [],
...
}